% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{angle_to_colour}
\alias{angle_to_colour}
\title{Convert Angle into colour code}
\usage{
angle_to_colour(x, alpha = 0.3)
}
\arguments{
\item{x}{Angle numeric vector}

\item{alpha}{A numeric vector between 0 and 1 for transparency}
}
\value{
a character vector for colour code
}
\description{
Convert Angle into colour code
}
\examples{
\dontrun{
col  <-  angle_to_colour(x)
}

}
\keyword{internal}
