% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyCCA.R
\name{BiplotCCA}
\alias{BiplotCCA}
\title{Generate CCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)}
\usage{
BiplotCCA(
  x,
  ncol = 2,
  nrow = 1,
  jpeg = FALSE,
  select = FALSE,
  colour = "Spectral",
  xlim = NULL,
  ylim = NULL,
  max_cells_displayed = 30000,
  verbose = FALSE
)
}
\arguments{
\item{x}{A TockyPrepData object produced by the function Tocky}

\item{ncol}{The number of columns in plot}

\item{nrow}{The number of rows in plot}

\item{jpeg}{A logical argument. If FALSE, it will open a device window in which plots are generated.}

\item{select}{Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.}

\item{colour}{Either 'Spectral' or 'BlueRed' for Angle colour key.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}

\item{max_cells_displayed}{The maximum number of cells to display in the plots}

\item{verbose}{Logical indicating whether to print progress messages and outputs.}
}
\value{
Generates a CCA heatmap plot and optionally saves it as a jpeg file.
}
\description{
Generate CCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)
}
\examples{
\dontrun{
BiplotCCA(x)
}
}
