% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{cluster_cell_num}
\alias{cluster_cell_num}
\title{Calculate the number of cells in each cluster in each sample using a Tocky Object.}
\usage{
cluster_cell_num(x)
}
\arguments{
\item{x}{A TockyPrepData after running the function TockyClustering.}
}
\value{
The slot Reduction will contain the new slot cluster_cellnumbers, which includes the cell number of each cluster in each sample
}
\description{
Calculate the number of cells in each cluster in each sample using a Tocky Object.
}
\examples{
\dontrun{
x <- cluster_cell_num(x)
}

}
\keyword{internal}
