% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{color_palette}
\alias{color_palette}
\title{To produce distinct colour codes}
\usage{
color_palette(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
a character vector for colour code
}
\description{
To produce distinct colour codes
}
\examples{
\dontrun{
col  <-  color_palette(x)
}
}
\keyword{internal}
