% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfinding.R
\name{DijkstraTockyPath}
\alias{DijkstraTockyPath}
\title{Dijkstra-Tocky Pathfinding}
\usage{
DijkstraTockyPath(x, origin_node, destination_node, verbose = TRUE)
}
\arguments{
\item{x}{A TockyPrepData containing the network and relevant data.}

\item{origin_node}{The starting node (cluster) for the path.}

\item{destination_node}{The ending node (cluster) for the path.}

\item{verbose}{Logical indicating whether to print progress messages and outputs.}
}
\value{
The original TockyPrepData is returned with an additional list attached
        containing the ordered paths and the closest path.
}
\description{
This function identifies paths within a Tocky Network using Dijkstra-Tocky Pathfinding
}
\details{
The function computes mean TimerAngles for each cluster, identifies
         all possible increasing paths from the origin to the destination,
         and then calculates the total weight for each path to find the closest one.
         Paths are determined by increasing TimerAngle values, ensuring that each
         step along a path moves to a node with a higher TimerAngle.
}
\examples{
\dontrun{
x <- DijkstraTockyPath(x, origin_node = '3', destination_node = '21')
}
}
