% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{plot_color_code}
\alias{plot_color_code}
\title{To produce colour key as a gradient rectangle}
\usage{
plot_color_code(
  expression = NULL,
  x,
  y,
  width,
  height,
  colour = "Spectral",
  method = "Expression"
)
}
\arguments{
\item{expression}{A numeric vector defining the expression to be plotted.}

\item{x}{A number. The x position of the lower left corner of the rectangle}

\item{y}{A number. The y position of the lower left corner of the rectangle}

\item{width}{A number. The width of the rectangle}

\item{height}{A number. The height of the rectangle}

\item{colour}{Either 'Spectral' or 'BlueRed' for Angle colour key.}

\item{method}{Either 'Angle' or 'Spectral' for Angle and any expression, respectively.}
}
\description{
To produce colour key as a gradient rectangle
}
\examples{
\dontrun{
plot(1:5)
plot_color_code(x = 3, y = 4.5, width = 1, height = 0.3, colour = 'Spectral', method = 'Angle')
}
}
