% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{plot_network_path}
\alias{plot_network_path}
\title{Plot Network Path}
\usage{
plot_network_path(
  network,
  path = NULL,
  vertex_col = "grey",
  title = "Angle Path"
)
}
\arguments{
\item{network}{An igraph network object to be visualized. The function expects this object to represent a graph with nodes and edges properly defined.}

\item{path}{Optional; a vector of node names indicating the sequence of nodes along the path to be highlighted. If provided, this path will be visually distinguished in the plot.}

\item{vertex_col}{A character string indicating the default color for vertices not in the path. This allows for customization of the network's appearance.}

\item{title}{A character string providing a title for the network plot. This allows for basic customization of the plot's appearance.}
}
\value{
Invisibly returns NULL. The function is used for its side effect of plotting the network.
}
\description{
Visualizes a network using the Fruchterman-Reingold layout, with the option to highlight a specific path and customize vertex colors.
Nodes and edges along the path are highlighted to distinguish them from the rest of the network.
}
\examples{
\dontrun{
plot_network_path(net, path = c("node1", "node2", "node3"),
vertex_col = "lightblue", title = "Sample Network Path")
}
}
\keyword{internal}
