% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPercentageClusters.R
\name{PlotClusterPercentage}
\alias{PlotClusterPercentage}
\title{Plot Cluster Percentage by Bar plot}
\usage{
PlotClusterPercentage(
  x,
  p_adjust_method = "fdr",
  colours = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A TockyPrepData after running the function ClusteringPCA.}

\item{p_adjust_method}{A method for p-value adjustment in statistical tests.}

\item{colours}{(optional) A vector specifying colours for different groups in the plot. For example, colours = c("purple", "black").}

\item{verbose}{Logical indicating whether to print progress messages and outputs.}
}
\value{
The TockyPrepData with added statistics and plots for significant clusters.
}
\description{
Plot Cluster Percentage by Bar plot
}
\examples{
\dontrun{
x <- PlotClusterPercentage(x)
}
}
