% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyPCA.R
\name{PlotPCAHeatmap}
\alias{PlotPCAHeatmap}
\title{Generate PCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)}
\usage{
PlotPCAHeatmap(
  x,
  ncol = 2,
  nrow = 1,
  jpeg = FALSE,
  select = FALSE,
  biplot_scaling = 3,
  colour = "Spectral",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{A TockyPrepData object produced by the function Tocky}

\item{ncol}{The number of columns in plot}

\item{nrow}{The number of rows in plot}

\item{jpeg}{A logical argument. If FALSE, it will open a device window in which plots are generated.}

\item{select}{Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.}

\item{biplot_scaling}{A number for multiplying biplot values for visibility. Default is 3.}

\item{colour}{Either 'Spectral' or 'BlueRed' for Angle colour key.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}
}
\value{
Generates a CCA heatmap plot and optionally saves it as a jpeg file.
}
\description{
Generate PCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)
}
\examples{
\dontrun{
PlotPCAHeatmap(x)
}
}
