% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyCCA.R
\name{PlotTockyCCA}
\alias{PlotTockyCCA}
\title{Generate CCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)}
\usage{
PlotTockyCCA(x, ncol = 4, nrow = 3, jpeg = FALSE, select = FALSE)
}
\arguments{
\item{x}{A TockyPrepData object produced by the function Tocky}

\item{ncol}{The number of columns in plot}

\item{nrow}{The number of rows in plot}

\item{jpeg}{A logical arguement. If FALSE, it will open a device window in which plots are generated.}

\item{select}{Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.}
}
\description{
Generate CCA heatmap plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)
}
\examples{
\dontrun{
PlotTockyCCA(x)
}
}
