% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyNetwork.R
\name{PlotTockyNetwork}
\alias{PlotTockyNetwork}
\title{Plot A Tocky Network}
\usage{
PlotTockyNetwork(
  x,
  reduction = "CCA",
  select_variable = FALSE,
  edge_scale = NULL,
  mds = FALSE,
  reflect = NULL,
  log2fc = FALSE,
  p_adjust = NULL
)
}
\arguments{
\item{x}{A TockyPrepData after running the function TockyNetwork.}

\item{reduction}{Choose whether to use PCA or CCA as a reduction method.}

\item{select_variable}{Whether to choose a variable (marker expression). The default is FALSE and produces Tocky Angle and Intensity.}

\item{edge_scale}{A scale factor for edge thickness.}

\item{mds}{If Multidimensional Scaling is used for constructing network graph.}

\item{reflect}{If Multidimensional Scaling is used, graph can be refelected by either 'horizontal' or 'vertical'.}

\item{log2fc}{Logical. If TRUE, the colours of nodes are determined by log2 fold change between the two groups.}

\item{p_adjust}{If p_adjust is "none", p-value adjustment is not used.
Note that p-values from two-group comparisons are stored within the TockyPrepData object and have been calculated by 'PlotClusterPercentage'.}
}
\value{
Network plot is produced, showing clusters as vertices and distance between clusters as edges.
}
\description{
Plot A Tocky Network
}
\examples{
\dontrun{
PlotTockyNetwork(x)
}
}
