% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyPCA.R
\name{PlotTockyPCA}
\alias{PlotTockyPCA}
\title{Produce PCA plot}
\usage{
PlotTockyPCA(x, jpeg = FALSE, cluster = FALSE, filename = NULL)
}
\arguments{
\item{x}{A TockyPrepData after running the function Tocky.}

\item{jpeg}{A logical arguement. If FALSE, it will open a device window in which plots are generated.}

\item{cluster}{A logical arguement. If TRUE, clusters are coloured in output plots.}

\item{filename}{A character string for file name}
}
\value{
A TockyPrepData simply to avoid null return.
}
\description{
Produce PCA plot
}
\examples{
\dontrun{
PlotTockyPCA(x)
}
}
