% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfinding.R
\name{showDijkstraTockyPath}
\alias{showDijkstraTockyPath}
\title{Display Closest Tocky Angle Gradient Path}
\usage{
showDijkstraTockyPath(x, origin_node, destination_node)
}
\arguments{
\item{x}{A TockyPrepData that has already been processed with DijkstraTockyPath
to compute the Tocky Angle Gradient Path.}

\item{origin_node}{The starting node (cluster) for the path.}

\item{destination_node}{The ending node (cluster) for the path.}
}
\value{
Prints the closest path and also returns it for further analysis or usage.
}
\description{
Retrieves and displays the closest Tocky Angle Gradient Path from a TockyPrepData.
This function is intended for use after running DijkstraTockyPath to quickly
access and review the computed closest path based on increasing TimerAngle values.
}
\examples{
\dontrun{
showDijkstraTockyPath(x)
}
}
