% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyCCA.R
\name{TockyCCA}
\alias{TockyCCA}
\title{Perform Dimensional Reduction using Canonical Correspondence Analysis (CCA) for TockyPrepData.}
\usage{
TockyCCA(x, variables = NULL, marker_neg_gate = TRUE, select = FALSE)
}
\arguments{
\item{x}{A TockyPrepData after running the function Tocky.}

\item{variables}{Variables (markers) for CCA analysis. If NULL, variables are to be chosed interactively.}

\item{marker_neg_gate}{Whether autofluorescence values are all considered to be zero or not. This approach is recommended. Perform DefineNegative for the same variables to be used by TockyCCA in advance.}

\item{select}{Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.}
}
\value{
The slot Reduction will contain CCA results.
}
\description{
Perform Dimensional Reduction using Canonical Correspondence Analysis (CCA) for TockyPrepData.
}
\examples{
\dontrun{
x <- TockyCCA(x)
}
}
