% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyClustering.R
\name{TockyClustering}
\alias{TockyClustering}
\title{Clustering Cells Using Dimensional Reduction by TockyPCA.}
\usage{
TockyClustering(
  x,
  choose_dimension = FALSE,
  max_cells_displayed = 30000,
  num_centre = FALSE
)
}
\arguments{
\item{x}{A TockyPrepData after running the function TockyPCA.}

\item{choose_dimension}{Whether the number of dimension is specified by PCA plot. Default is FALSE.}

\item{max_cells_displayed}{The number of cells displayed in plots for interatcive PCA plots.}

\item{num_centre}{Whether the number of clusters is specified. Default is FALSE and the number of variables will be used.}
}
\value{
The slot Reduction will contain the new slot PCAclusters, which includes kmeans clustering result
}
\description{
Clustering Cells Using Dimensional Reduction by TockyPCA.
}
\examples{
\dontrun{
x <- TockyClustering(x)
}
}
