% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyNetwork.R
\name{TockyNetwork}
\alias{TockyNetwork}
\title{Produce A Network of Tocky Clusters}
\usage{
TockyNetwork(x, reduction = "CCA", cut_off = 0.2)
}
\arguments{
\item{x}{A TockyPrepData after running the function TockyClustering.}

\item{reduction}{Choose whether to use PCA or CCA as a reduction method. When using CCA, TockyCCA must have been applied to the TockyPrepData.}

\item{cut_off}{Threshold value as a quantile percentage for edge connection. For example, the default 0.2 will set the threshold at 20\% quantile of distance between clusters, connecting neighbor clusters.}
}
\value{
A revised TockyPrepData containing an igraph network object. Network is constructed based on the neighbor proximity.
}
\description{
Produce A Network of Tocky Clusters
}
\examples{
\dontrun{
x <- TockyNetwork(x)
}
}
