% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyPCA.R
\name{TockyPCA}
\alias{TockyPCA}
\title{Perform Principal Component Analysis for TockyPrepData.}
\usage{
TockyPCA(
  x,
  variables = NULL,
  marker_neg_gate = TRUE,
  cleaning = FALSE,
  select = TRUE,
  Timer = FALSE
)
}
\arguments{
\item{x}{A TockyPrepData after running the function Tocky.}

\item{variables}{Variables (markers) for PCA analysis. If NULL, variables are to be chosed interactively.}

\item{marker_neg_gate}{Whether autofluorescence values are all considered to be zero. This approach is recommended and requires DefineNegative.}

\item{cleaning}{Whether data cleaning is performed to remove cells with a zero value for a marker. This is not recommended unless you have a reason that you cannot use DefineNegative to collapse negative data.}

\item{select}{Whether to interactively select markers to be processed. If FALSE, all the log-transformed markers apart from Timer fluorescence (stored in the TockyPrepData) will be used.}

\item{Timer}{Whether Timer data (normalised or Angle/Intensity) are used.}
}
\description{
Perform Principal Component Analysis for TockyPrepData.
}
\examples{
\dontrun{
x <- TockyPCA(x, variables = NULL, cleaning = FALSE, marker_neg_gate = TRUE)
}
}
